% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTime.R
\name{analyzeTime}
\alias{analyzeTime}
\title{Time Analysis}
\usage{
analyzeTime(title, fill = c("cornflowerblue", "#009E73"), func1, func2)
}
\arguments{
\item{title}{title of the graph to be created}

\item{fill}{a vector of length two of colors for the bargraphs}

\item{func1}{first function to be analyzed}

\item{func2}{second function to be analyzed}
}
\value{
a bar graph displaying the user time comparison of the two functions
}
\description{
\code{analyzeTime} compares the time it takes to run two functions
}
\details{
This function takes two functions and displays the user time each
took, as well as a bar graph of the comparison
}
\examples{

analyzeTime(title = "Fibonacci vs. Memoized Fibonacci", func1 = fib(30), func2 = fib_memoized(30,createList(30)))
}
