% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createList.R
\name{createList}
\alias{createList}
\title{Create list for memoized fibonacci}
\usage{
createList(n)
}
\arguments{
\item{n}{an integer}
}
\value{
a n-length list with 1 as the first two elements, and -1 for the rest
}
\description{
\code{createList} creates a list of length n for memoized fibonacci
}
\examples{
createList(10)
}
