% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib.R
\name{fib}
\alias{fib}
\title{Basic Fibonacci}
\usage{
fib(n)
}
\arguments{
\item{n}{an integer}
}
\value{
the nth fibonacci number
}
\description{
\code{fib} computes the nth fibonacci number (1-indexed)
}
\details{
This function calls itself for each recursive computation.
}
\examples{
fib(10)
}
