% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib_memoized.R
\name{fib_memoized}
\alias{fib_memoized}
\title{Memoized Fibonacci}
\usage{
fib_memoized(n, newlist)
}
\arguments{
\item{n}{an integer}

\item{newlist}{createList(n)}
}
\value{
the nth fibonacci number
}
\description{
\code{fib_memoized} computes the nth fibonacci number (1-indexed)
}
\details{
This function uses memoization and is more efficient than \code{fib}
}
\examples{
fib_memoized(10,createList(10))[[1]]
}
