% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiply By Recursive Addition}
\usage{
mult(x, y)
}
\arguments{
\item{x}{first value for multiplication}

\item{y}{second value for multiplication}
}
\value{
the result of x multiplied by y
}
\description{
\code{mult} multiplies the first and second parameter
and returns result
}
\details{
This function does the same operation as * but by recursively adding values
}
\examples{

mult(2,3)
}
