% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Density Plot of Recusive calls}
\usage{
plotDensity(data, title = "Title", color = "cornflowerblue",
  border = "red")
}
\arguments{
\item{data}{output from recDist}

\item{title}{title of graph (defaults to "Title")}

\item{color}{color of area of density plot (defaults to "cornflowerblue")}

\item{border}{color of outline of density plot (defaults to "red")}
}
\value{
density plot of data
}
\description{
\code{plotDensity} Plots a density plot of the recursive calls
}
\details{
This function takes output from recDist and plots a density plot of all recursive calls
}
\examples{
capture.output(recursiveCalls(fib,10), file = "output.txt")
fibResult <- saveCalls("output.txt")
plotDensity(recDist(fibResult))
}
