% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Distribution Plot}
\usage{
plotDist(data, xlabel = "x label", ylabel = "y label",
  title = "Title", color = "darkblue")
}
\arguments{
\item{data}{result of recDist}

\item{xlabel}{x label title of graph (defaults to "x label")}

\item{ylabel}{y label title of graph (defaults to "y label")}

\item{title}{title of graph (defaults to "Title")}

\item{color}{color of points (defaults to "darkblue")}
}
\value{
the result of x multiplied by y
}
\description{
\code{plotDist} Plots a distribution plot of the result of recDist.
}
\details{
This function is a visualization tools for the result of recDist. It allows the user
to see how often subproblems are called.
}
\examples{
capture.output(recursiveCalls(fib,10), file = "output.txt")
fibResult <- saveCalls("output.txt")
plotDist(recDist(fibResult))
}
