% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recDist.R
\name{recDist}
\alias{recDist}
\title{Count Recursive Calls}
\usage{
recDist(result)
}
\arguments{
\item{result}{resulting character vector from saveCalls}
}
\value{
dataframe containing recursive subproblems in sorted order by frequency
}
\description{
\code{recDist} Counts all subproblems in order of frequency of times called and
returns the frequency table as a dataframe.
}
\details{
Takes the result of saveCalls and gives a frequency table sorted by frequency of recursive subproblem.
}
\examples{

fibResult <- saveCalls("output.txt")
recDist(fibResult)
}
