% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveCalls.R
\name{recursiveCalls}
\alias{recursiveCalls}
\title{Check All Recursive Calls}
\usage{
recursiveCalls(recFunc, ...)
}
\arguments{
\item{...}{the parameters of the function to be called}

\item{recfunc:}{the recursive function to be called without the parenthesis or the arguments;
must have at least one recursive function call}
}
\value{
the result of calling the input function with the given parameters
}
\description{
\code{recursiveCalls} computes the result of a recursive function
call and prints all recursive calls
}
\details{
This function takes a recursive function and its parameters and returns the
result of that function call and also prints all recursive calls. Useful for
checking for errors in recursive functions.
}
\examples{

recursiveCalls(fib,5)
}
