% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCalls.R
\name{saveCalls}
\alias{saveCalls}
\title{Save All Recursive Calls}
\usage{
saveCalls(filename)
}
\arguments{
\item{filename}{the file you saved in capture.output}
}
\value{
Character vector containing every recursive call
}
\description{
\code{saveCalls} Returns character vector containing all recursive subproblems
}
\details{
Use this function in conjunction with capture.output and recursiveCalls.
Call capture.output where first parameter is calling recursiveCalls on any recursive function
and second parameter is the file name you would like to output it to.
}
\examples{

capture.output(recursiveCalls(fib,10), file = "output.txt")
fibResult <- saveCalls("output.txt")


newlist1 <- createList(10)
capture.output(recursiveCalls(fib_memoized,10,newlist1), file = "output.txt")
memoizedResult <- saveCalls("output.txt")
}
