% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{saveOutput}
\alias{saveOutput}
\title{Save Output (Printed Results) from a function}
\usage{
saveOutput(..., file = NULL, append = FALSE, type = c("output",
  "message"), split = FALSE)
}
\arguments{
\item{x:}{first value for multiplication}

\item{y:}{second value for multiplication}
}
\value{
the result of x multiplied by y
}
\description{
\code{saveOutput} Takes print function calls from function and returns output
in desired format.
}
\details{
This function takes every print call in the function called as a parameter
and adds the argument of the print function to a character vector and returns
the output in specified file format.
}
\examples{

mult(2,3)
}
